Use ParallelTestRunner and split test files for parallelism (#254)
