+++ Project.toml
@@ -20,10 +20,10 @@ Statistics = "1"
 julia = "1.10"
 
 [extras]
+ParallelTestRunner = "d3525ed8-44d0-4b2c-a655-542cee43accc"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
-TimerOutputs = "a759f4b9-e2f1-59dc-863e-4aeb61b1ea8f"
 Unitful = "1986cc42-f94f-5a68-af5c-568840ba703d"
 
 [targets]
-test = ["Random", "Test", "TimerOutputs", "Unitful"]
+test = ["ParallelTestRunner", "Random", "Test", "Unitful"]
