+++ test/mixed_tensors.jl
@@ -1,5 +1,5 @@
-@testsection "MixedTensors" begin
-    @testsection "regular_conversions" begin
+@testset "MixedTensors" begin
+    @testset "regular_conversions" begin
         @test Tensors.isregular(MixedTensor{1, Tuple{3}}) # order 1 always regular
         for order in 2:4
             dims = rand(1:2, order)
@@ -26,8 +26,8 @@
             @test (@inferred Tensors.makemixed(tt))::TR === tr
         end
     end
-    @testsection "construction" begin
-        @testsection "from tuples" begin
+    @testset "construction" begin
+        @testset "from tuples" begin
             data = ntuple(_ -> rand(), 6)
             TD = ForwardDiff.Dual{Nothing, Float64, 3}
             TB = MixedTensor2{2, 3}
@@ -39,7 +39,7 @@
             # Heterogeneous tuple
             @test isa(TB((1.f0, 1.0, 1, zero(TD), 2, 3)), MixedTensor2{2, 3, TD, 6})
         end
-        @testsection "from function" begin
+        @testset "from function" begin
             T2 = MixedTensor{2, Tuple{(rand(1:3, 2)...)}}
             T3 = MixedTensor{3, Tuple{(rand(1:3, 3)...)}}
             T4 = MixedTensor{4, Tuple{(rand(1:3, 4)...)}}
@@ -48,7 +48,7 @@
             @test T3((i, j, k) -> m3[i, j, k]) == m3
             @test T4((i, j, k, l) -> m4[i, j, k, l]) == m4 
         end
-        @testsection "from slice" begin
+        @testset "from slice" begin
             v = rand(Vec{2})
             t = rand(Tensor{2,2})
             m = MixedTensor2{2, 3}((v.data..., t.data...))
@@ -60,27 +60,27 @@
             @test m[2, :] == Vec((v[2], t[2,1], t[2,2]))
         end
     end
-    @testsection "basic ops" begin
+    @testset "basic ops" begin
         m23 = rand(MixedTensor2{2, 3})::MixedTensor2{2, 3}
         m21 = rand(MixedTensor2{2, 1})::MixedTensor2{2, 1}
         m231 = rand(MixedTensor3{2, 3, 1})::MixedTensor3{2, 3, 1}
         m2332 = rand(MixedTensor4{2, 3, 3, 2})::MixedTensor4{2, 3, 3, 2}
         m2131 = rand(MixedTensor4{2, 1, 3, 1})::MixedTensor4{2, 1, 3, 1}
-        @testsection "transpose" begin
+        @testset "transpose" begin
             @test ((@inferred transpose(m23))::MixedTensor2{3, 2})[2, 1] == m23[1, 2]
             @test ((@inferred transpose(m21))::MixedTensor2{1, 2})[1, 2] == m21[2, 1]
             @test transpose(m23) == m23'
             @test ((@inferred majortranspose(m2332))::MixedTensor4{3, 2, 2, 3})[3, 2, 2, 1] == m2332[2, 1, 3, 2]
             @test ((@inferred minortranspose(m2131))::MixedTensor4{1, 2, 1, 3})[1, 2, 1, 3] == m2131[2, 1, 3, 1]
         end
-        @testsection "norm" begin
+        @testset "norm" begin
             for order in (1, 2, 3, 4)
                 t = rand(MixedTensor{order, Tuple{(rand(1:3, order)...)}})
                 @test norm(t) ≈ norm(collect(t.data))
             end
         end
     end
-    @testsection "basic arithmetric" begin
+    @testset "basic arithmetric" begin
         for T in (Float32, Float64, ForwardDiff.Dual{Nothing, Float64, 2})
             m23_f64 = rand(MixedTensor2{3, 2})
             m23_T = rand(MixedTensor2{3, 2, T})
@@ -97,7 +97,7 @@
             end
         end
     end
-    @testsection "dot" begin
+    @testset "dot" begin
         function dot_index(a::MixedTensor2{di, ds}, v::Union{Vec{ds}, MixedTensor{1, Tuple{ds}}}) where {di, ds}
             return Vec{di}(i -> sum(a[i, j] * v[j] for j in 1:ds))
         end
@@ -123,7 +123,7 @@
         @test (@inferred m32 ⋅ v)::Vec{3} ≈ m32 ⋅ m2
     end
 
-    @testsection "dcontract" begin
+    @testset "dcontract" begin
         function dcontract_index(a::MixedTensor2{di, dj}, b::MixedTensor2{di, dj}) where {di, dj}
             return sum(a.data[i] * b.data[i] for i in 1:(di*dj))
         end
@@ -160,7 +160,7 @@
         @test ((@inferred m3322 ⊡ t22)::Tensor{2, 3}) ≈ Tensors.regular_if_possible(dcontract_index(m3322, t22))
     end
 
-    @testsection "otimes" begin
+    @testset "otimes" begin
         otimes_index(u::Vec{di}, v::Vec{dj}) where {di, dj} = MixedTensor2{di, dj}((i,j) -> u[i] * v[j])
         otimes_index(u::MixedTensor2{di, dj}, v::Vec{dk}) where {di, dj, dk} = MixedTensor3{di, dj, dk}((i,j,k) -> u[i, j] * v[k])
         otimes_index(u::Vec{di}, v::MixedTensor2{dj, dk}) where {di, dj, dk} = MixedTensor3{di, dj, dk}((i,j,k) -> u[i] * v[j, k])
@@ -176,7 +176,7 @@
         @test m31 ⊗ m23 ≈ otimes_index(m31, m23)
     end
 
-    @testsection "Automatic Differentation" begin
+    @testset "Automatic Differentation" begin
         # As code is mostly shared with regular `Tensor`s, the main part 
         # here is to test the extraction and insertion parts in the right locations. 
         # So we make some artifical tests but don't need too much nested functions.
@@ -194,7 +194,7 @@
         dfdx = MixedTensor2{2, 3}((1.0, v[3], 1.0, 0.0, 0.0, v[1]))
         @test gradient(foo, v) ≈ dfdx
     end
-    @testsection "errors" begin
+    @testset "errors" begin
         m32 = rand(MixedTensor2{2,3})
         @test_throws ArgumentError m32^2
     end
